==> 1_Participants.Rmd <==
---
title: "Participants"
output: 
  html_document
---

<!-- Set general settings -->

```{r setup, message = FALSE, warning = FALSE}

# Set general settings for markdown file
knitr::opts_chunk$set(
  message = FALSE,
  warning = FALSE)


# Clear environment
rm(list = ls())


# Load packages
library(dplyr)      # for data wrangling
library(knitr)      # for integrating computing and reporting in markdown
library(kableExtra) # for customizing appearance of tables
library(foreign)    # for reading in SPSS file
library(psych)      # for describeBy function
library(tidyr)      # for gather and spread function


# Load functions
source("./functions/my_table_template.R")


# Turn off scientific notation
options(scipen = 999)
```
<br>

## Sample Description
***

* The final sample consisted of 28 patients with OCD and 28 healthy control (HC) participants. 

* Patients with OCD and healthy control participants were individually matched for age, gender, and educational level.

* Patients were diagnosed using the Structured Clinical Interview for DSM-IV (SCID-I; Wittchen et al., 1997) and were recruited from the outpatient clinic for OCD at Humboldt-Universität zu Berlin. 

* All participants were native German speakers and had normal or corrected-to-normal vision. 

* Exclusion criteria were lifetime diagnosis of bipolar, psychotic, or substance-related disorders; neurological diseases; and use of neuroleptic medication in the last 3 months or benzodiazepines in the last week. 

* Additional exclusion criteria for control participants were any current or past psychiatric disorder or psychotherapeutic treatment.  
<br><br>

## Sample Characteristics
***

This table corresponds to Table 1 in the manuscript.

```{r demographics}

# Load data
participant_data <- read.spss("./data/Probandendaten.sav", to.data.frame = TRUE)
load(file = "./data/Trait_Data.rda")


# Prepare dataframe
demographics <- participant_data %>%
  # Prepare id variable for merging and add group variable
  dplyr::mutate(participant_id = as.factor(paste0(substr(Code, 13, 13), "_", substr(Code, 14, 15))),
                group = as.factor(ifelse(substr(participant_id, 1, 1) == "C", "HC", "OCD"))) %>%
  # Exclude P_02 and C_02 and row with NA
  dplyr::filter(participant_id != "P_02" & participant_id != "C_02" & !is.na(Alter)) %>%
  # Rename columns
  dplyr::rename(
    "Age (years)"                = Alter,
    "Gender"                     = Geschlecht,
    "Years of education"         = Schuljahre,
    "Y-BOCS total score"         = YBOCS_gesamt,
    "Y-BOCS obsessions"          = YBOCS_Gedanken,
    "Y-BOCS compulsions"         = YBOCS_Handlungen,
    "Verbal intelligence (WST)"  = WST,
    "Handedness"                 = "Händigkeit") %>%
  # Merge trait data
  dplyr::left_join(., traits, by = "participant_id") %>%
  # Select and reorder columns
  dplyr::select(c("group", "Age (years)", "Gender", "Years of education", "Verbal intelligence (WST)", "Handedness",
    "BDI-II",  "OCI-R", "PSWQ", "STAI trait", "Y-BOCS total score",  "Y-BOCS obsessions", "Y-BOCS compulsions"))


# Transform WST scores (range 24-37) to IQ scores (according to WST manual)
demographics[demographics$"Verbal intelligence (WST)" == 24, ]$"Verbal intelligence (WST)" <- 89
demographics[demographics$"Verbal intelligence (WST)" == 25, ]$"Verbal intelligence (WST)" <- 90
demographics[demographics$"Verbal intelligence (WST)" == 27, ]$"Verbal intelligence (WST)" <- 93
demographics[demographics$"Verbal intelligence (WST)" == 28, ]$"Verbal intelligence (WST)" <- 95
demographics[demographics$"Verbal intelligence (WST)" == 29, ]$"Verbal intelligence (WST)" <- 97
demographics[demographics$"Verbal intelligence (WST)" == 30, ]$"Verbal intelligence (WST)" <- 99
demographics[demographics$"Verbal intelligence (WST)" == 31, ]$"Verbal intelligence (WST)" <- 101
demographics[demographics$"Verbal intelligence (WST)" == 32, ]$"Verbal intelligence (WST)" <- 104
demographics[demographics$"Verbal intelligence (WST)" == 33, ]$"Verbal intelligence (WST)" <- 107
demographics[demographics$"Verbal intelligence (WST)" == 34, ]$"Verbal intelligence (WST)" <- 110
demographics[demographics$"Verbal intelligence (WST)" == 35, ]$"Verbal intelligence (WST)" <- 114
demographics[demographics$"Verbal intelligence (WST)" == 36, ]$"Verbal intelligence (WST)" <- 118
demographics[demographics$"Verbal intelligence (WST)" == 37, ]$"Verbal intelligence (WST)" <- 122


# Rescore one handedness score (according to EHI score that deviates from self report)
demographics[demographics$"Handedness" == "beidhändig" & 
             demographics$"Age (years)" == 29, ]$"Handedness" <- "rechts"



### Numeric variables

# Extract demographics per group
demographics_groups <- list()
demographics_groups <- describeBy(demographics %>% dplyr::select(-c("Gender", "Handedness", "group")), demographics$group)
ocd <- as.data.frame(demographics_groups$`OCD`)
hc  <- as.data.frame(demographics_groups$`HC`)


# Create table with demographics formatted as "M (SD)"
demographics_groups <- as.data.frame(cbind(
  Characteristic = rownames(ocd),
  ocd = paste0(format(round(ocd$mean, 2), nsmall = 2), " (", format(round(ocd$sd, 2), nsmall = 2), ")"),
  hc  = paste0(format(round(hc$mean,  2), nsmall = 2), " (", format(round(hc$sd,  2), nsmall = 2), ")")))


# Calculate t tests
t_tests <- demographics %>%
  dplyr::select("Age (years)", "Years of education", "Verbal intelligence (WST)", "BDI-II", "OCI-R",
    "PSWQ", "STAI trait", "group") %>%
  tidyr::gather(key = Characteristic, value = value, -group) %>%
  dplyr::group_by(group, Characteristic) %>%
  dplyr::summarise(value = list(value)) %>%
  tidyr::spread(group, value) %>%
  dplyr::ungroup() %>%
  dplyr::group_by(Characteristic) %>%
  dplyr::mutate("Test statistic" = t.test(unlist(HC), unlist(OCD))$statistic,
                              df = t.test(unlist(HC), unlist(OCD))$parameter,
                              p  = t.test(unlist(HC), unlist(OCD))$p.value) %>%
  dplyr::select(-c(HC, OCD))  %>%
  dplyr::ungroup()


# Combine demographics and t test results
demographics_numeric <- left_join(demographics_groups, t_tests, by = "Characteristic")



### Count variables (gender, handedness)

# Gender: calculate chi-squared test and extract counts formatted as "female:male"
gender_test <- chisq.test(demographics$Gender, demographics$group)
gender      <- demographics %>% dplyr::group_by(group) %>% dplyr::count(Gender)
gender_ocd  <- paste0(gender[gender$group == "OCD" & gender$Gender == "weiblich", ]$n, ":",
                      gender[gender$group == "OCD" & gender$Gender == "männlich", ]$n)
gender_hc   <- paste0(gender[gender$group == "HC"  & gender$Gender == "weiblich", ]$n, ":",
                      gender[gender$group == "HC"  & gender$Gender == "männlich", ]$n)


# Handedness: calculate Fishers's exact test and extract counts formatted as "right:left:ambidextrous"
hand_test <- fisher.test(demographics$Handedness, demographics$group)
hand      <- demographics %>% dplyr::group_by(group) %>% dplyr::count(Handedness)
hand_ocd  <- paste0(hand[hand$group == "OCD" & hand$Handedness == "rechts", ]$n, ":",
                    hand[hand$group == "OCD" & hand$Handedness == "links", ]$n, ":",
                    hand[hand$group == "OCD" & hand$Handedness == "beidhändig", ]$n)
hand_hc   <- paste0(hand[hand$group == "HC"  & hand$Handedness == "rechts", ]$n, ":",
                    hand[hand$group == "HC"  & hand$Handedness == "links", ]$n, ":",
                    length(hand[hand$group == "HC" & hand$Handedness == "beidhändig", ]$n))


# Combine demographics and chi-squared test/Fisher's exact test
demographics_factorial <- as.data.frame(cbind(
  Characteristic = c("Gender (n female:male)", "Handedness (n right:left:ambidextrous)"),
  ocd = c(gender_ocd, hand_ocd),
  hc  = c(gender_hc,  hand_hc),
  "Test statistic" = c(gender_test$statistic, "-"),
  df = c(gender_test$parameter, "-"),
  p = c(gender_test$p.value, hand_test$p.value)))



### Create and display table

# Re-order rows and format p values
table <- rbind(demographics_numeric[c(1), ], demographics_factorial,
               demographics_numeric[c(2:nrow(demographics_numeric)), ])
table[, c("df", "Test statistic", "p")] <- as.numeric(unlist(table[, c("df", "Test statistic", "p")]))
table[, c("p")] <- format.pval(table[, c("p")], eps = 0.001, digits = 3)
table <- table %>% dplyr::rename("Patients with OCD" = ocd, "Healthy control participants" = hc)


# Display table
my_table_template(table, 
  caption = "Demographic and Clinical Characteristics") %>%
  add_footnote(
  "\n Note. Means are reported with standard deviations in parentheses (except for gender and handedness). 
  Welch’s t test was used for continuous variables. Age range was 18 to 55 years  (control participants: 20 to 54 years; 
  patients with OCD: 18 to 55 years). Handedness was assessed with the Edinburgh Handedness Inventory (Oldfield, 1971). 
  For Fisher’s exact test, there is no test statistic to report. Years of education include primary and secondary education. 
  WST = Wortschatztest; BDI-II = Beck Depression Inventory-II; OCI-R = Obsessive-Compulsive Inventory-Revised; 
  PSWQ = Penn State Worry Questionnaire; STAI = State-Trait Anxiety Inventory; Y-BOCS = Yale-Brown Obsessive 
  Compulsive Scale.", notation = "none")
```
<br><br>

## References
***
Wittchen, H., Zaudig, M., & Fydrich, T. (1997). *Strukturiertes klinisches Interview für DSM-IV (SKID).* Hogrefe. 	
<br><br><br>

## Session Info
***
```{r session-info}

sessionInfo()
```
